pyproject.toml
[project]
name = "cvpr22-fact"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "pandas>=2.2.3",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.7.2",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
    "tqdm>=4.67.1",
]
